Moved important playback statistics next to the seeker.
