+++ src/components/VideoControls/VideoControls.tsx
@@ -1,10 +1,11 @@
-import CurrentTimeLabel from "../CurrentTimeLabel";
 import MuteButton from "../MuteButton";
 import PlayButton from "../Play";
 import PlaybackRate from "../PlaybackRate";
-import React from "react";
+import React, { useState, useEffect } from "react";
 import Seeker from "../Seeker";
 import { makeStyles } from "@material-ui/core";
+import { getTime } from "../../utils";
+import Label from "../Label";
 
 export interface IVideoControls {
   duration: number;
@@ -19,7 +20,7 @@ export interface IVideoControls {
   watchStartTime: number;
 }
 
-const height = 76;
+const height = 90;
 const padding = 16;
 export const VIDEO_CONTROLS_HEIGHT = `${padding * 2 + height}px`;
 
@@ -32,7 +33,10 @@ const useStyles = makeStyles({
     padding,
   },
   seeker: {
-    width: `calc(100vw - ${padding * 2}px - 10px)`,
+    alignItems: "center",
+    display: "flex",
+    flexDirection: "row",
+    justifyContent: "space-between",
   },
   interactiveRoot: {
     display: "flex",
@@ -58,25 +62,53 @@ export default function VideoControls({
   watchStartTime,
 }: IVideoControls): JSX.Element {
   const styles = useStyles();
+
+  const [time, setTime] = useState("00:00:00");
+  const [now, setNow] = useState(0);
+  const [remainingTime, setRemainingTime] = useState("00:00:00");
+  const [remainingAtRate, setRemainingAtRate] = useState("00:00:00");
+
+  useEffect(() => {
+    setTime(getTime(currentTime));
+
+    const remaining = duration - currentTime;
+    setRemainingTime(getTime(remaining));
+
+    const rate = parseFloat(playbackRate) || 1;
+    setRemainingAtRate(getTime(remaining / rate));
+  }, [currentTime, duration, playbackRate, watchStartTime]);
+
+  useEffect(() => {
+    const interval = setInterval(() => {
+      setNow(Date.now());
+    }, 1000);
+    return () => clearInterval(interval);
+  }, [watchStartTime]);
+
   return (
     <div className={styles.root}>
-      <Seeker
-        className={styles.seeker}
-        currentTime={currentTime}
-        duration={duration}
-        updateCurrentTime={updateCurrentTime}
-      />
+      <div className={styles.seeker}>
+        <Label>{time}</Label>
+        <Seeker
+          currentTime={currentTime}
+          duration={duration}
+          updateCurrentTime={updateCurrentTime}
+        />
+        <Label>
+          -{remainingTime}/{getTime(duration)}
+        </Label>
+      </div>
 
       <div className={styles.interactiveRoot}>
         <div className={styles.interactiveRootLeft}>
           <PlayButton isPaused={isPaused} togglePaused={togglePaused} />
           <MuteButton isMuted={isMuted} toggleMuted={toggleMuted} />
-          <CurrentTimeLabel
+          {/* <CurrentTimeLabel
             currentTime={currentTime}
             duration={duration}
             playbackRate={playbackRate}
             watchStartTime={watchStartTime}
-          />
+          /> */}
         </div>
         <PlaybackRate value={playbackRate} updateValue={updatePlaybackRate} />
       </div>
